eclipse 3.3 **DEFAULT** formatter and clean up settings run on whole project. First attempt to solve formatting issues with Ciaran

+++ trunk/src/aima/learning/statistics/IrisDataSetNumerizer.java
@@ -66,7 +66,7 @@ public class IrisDataSetNumerizer implements Numerizer {
 		if (d > 1) {
 			return 1.0;
 		} else {
-			return (double) Math.round(d);
+			return Math.round(d);
 		}
 	}
